working on ubuntu 22.04 jammy port -- getting FreshVMs feature in bg-dev working * bug fix: setReturnValue: was failing if value started with a '-' * debugger: add * Try/Catch: improved --traceCatch mechanism * bgStackFreeze: added bgBASH_COMMAND to save BASH_COMMAND because stack traces are not only used in DEBUG scripts where BASH_COMMAND does not change * stringSplitIntoBashTokens: added function. this uses the history builtin and if finally a good way to parse bash tokens * bgtraceXTrace: made _bgtraceFile the default destination * bgExit: better interactive terminal detection so that it does not kill the user's terminal if there is one. * command_not_found_handle: small formatting change to error output

+++ core/bg_coreLibsMisc.sh
@@ -2022,7 +2022,7 @@ function whereAmIRunning()
 {
 	local _whereAmIRunningValue
 	local terminalBash; getTerminalPID terminalBash
-	if [ ${BASHPID:-0} -eq ${terminalBash:-0} ]; then
+	if [[ $- == *i* ]] || [ ${BASHPID:-0} -eq ${terminalBash:-0} ]; then
 		_whereAmIRunningValue="inTerminal"
 	elif [ ${BASHPID:-0} -eq $$ ]; then
 		_whereAmIRunningValue="inTopScript"
@@ -2070,8 +2070,18 @@ function bgexit()
 	local exitCode="$1"
 
 	# 2022-03 bobg: refactored this function completely. The previous version was working reasonably well for a long time but while
-	#               fixing command_not_found_handle, I begun to understand 'kill -TERM -$$' better. I thing this new version is
+	#               fixing command_not_found_handle, I begun to understand 'kill -TERM -$$' better. I think this new version is
 	#               more accurate and much easier to understand.
+	# 2022-08 bobg: the inTerminal detection did not handle the case of multiple nested interactive bashs like when a developer
+	#               runs 'bash' to get a prestine shell to do a test in.
+	#               I modifed whereAmIRunning to do better but this function still needs work. see TODO below
+	# TODO: the signature of bgExit needs to change to reflect two additional things to kill
+	#          1) background/siblings -- what if we are the backgroud. be more pgrpid aware in these calculations
+	#          2) parent pid up to the first interactive bash or the last one in this continuous chain.
+	#       note: that we should understand better the SIGTERM and SIGINT when sent to the whole pgrpid. maybe SIGINT blindly to the
+	#             whole pgrpid will work because interactive shells dont end with SIGINT
+	#       note: the inTerminal detection is probably fine now but the --complete is probably not b/c we need to walk up the parents
+	#             (and siblings) to decide what to kill
 
 	# TODO: consider merging this whereAmIRunning logic with the code in bg_importCntr.sh(lines ~300 to ~400)
 	local whereAmIRunning; whereAmIRunning "whereAmIRunning"
@@ -2886,7 +2896,7 @@ function command_not_found_handle()
 
 	local cmdName="$1"
 	local cmdline="$*"
-	local msg="Command not found. cmdline='$cmdline'"
+	local msg="bash: $cmdName: command not found"
 
 	# 2021-03 bobg: commented out this recursion check. from git commit - refined command_not_found_handle() to not throw assertError which was unable to end the script with bgExit --complete for some reason not yet undertood
 	# 2022-03 bobg: strange debugger bug - needs recursion check: second time launching, bgStackPrint and getUserCmpApp and others were not loaded which led to here and since this calls bgtraceStack led to inifinite recursion.
@@ -2909,11 +2919,11 @@ function command_not_found_handle()
 	fi
 
 	# write msg to stderr and to bgtrace (with a stack)
-	echo "$msg" >&2
 	if type -t bgtraceStack &>/dev/null && type -t bgStackPrint &>/dev/null; then
 		bgtraceStack
 		bgtraceIsActive && type -t bgStackPrint &>/dev/null && bgStackPrint
 	fi
+	echo "$msg" >&2
 	type -t bgtrace &>/dev/null && bgtrace "$msg"
 
 	# special case running in the terminal.
@@ -3253,9 +3263,9 @@ function assertError()
 			local throwingStatePID="$BASHPID"
 
 			if [ "$_ae_traceCatchFlag" ]; then
-				bgtrace "!!!throwing exception "
-				bgtrace "   PID of throw ='$throwingStatePID'"
-				bgtrace "   PID of catch ='$tryStatePID'"
+				bgtrace "!!!throwing exception that will be caught"
+				[ "$throwingStatePID" != "$$" ] && bgtrace "   PID of throw ='$throwingStatePID'"
+				[ "$throwingStatePID" != "tryStatePID" ] && bgtrace "   PID of catch ='$tryStatePID'"
 				bgtracePSTree
 				declare -g traceCatchFlag="1"
 			fi
@@ -3305,11 +3315,14 @@ function assertError()
 			# whose SIGUSR2 trap handler will install a DEBUG handler that will skip simple commands until it finds a "Catch".
 			# Some bash commands are interuptable by SIGINT but if we are running a bash script, all the parents between us and
 			# tryStatePID must be bash subshells stopped on bash functions.
-			[ "$_ae_traceCatchFlag" ] && bgtrace "!!! kill -SIGUSR2 $tryStatePID  "
+			[ "$traceCatchFlag" ] && {
+				bgtrace "!!! unwind is sending kill -SIGUSR2 $tryStatePID  "
+				bgDebuggerStepIntoPlumbing=1
+			}
 			kill -SIGUSR2 "$tryStatePID"     # this wont return if we are tryStatePID
 			(( ${#pidsToKill[@]} > 0 )) && kill -SIGINT "${pidsToKill[@]}"
-			[ "$tryStatePID" != "$BASHPID" ] && bgExit  ${_ae_exitCode:-36}
-			[ "$_ae_traceCatchFlag" ] && bgtrace "!!! returning normally. this is an error"
+			[ "$tryStatePID" != "$throwingStatePID" ] && bgExit  ${_ae_exitCode:-36}
+			[ "$traceCatchFlag" ] && bgtrace "!!! assertError: this exception is being caught but assertError is ending normally. this is an error"
 			;;
 
 		*)	echo "error: logic error. In assertError the action was computed to be '$tryStateAction' but should be one of catch,abort,continue,exitOneShell"
@@ -3363,14 +3376,6 @@ function Rethrow()
 			local tryStatePID="${bgBASH_tryStackPID[@]:0:1}"
 			local throwingStatePID="$BASHPID"
 
-			if [ "$_ae_traceCatchFlag" ]; then
-				bgtrace "!!!Re-throwing exception "
-				bgtrace "   PID of throw ='$throwingStatePID'"
-				bgtrace "   PID of catch ='$tryStatePID'"
-				bgtracePSTree
-				declare -g traceCatchFlag="1"
-			fi
-
 			## fill in pidsToKill with each pid between where the assert is being thrown (throwingStatePID) and where it will be
 			# caught (tryStatePID)
 			local pidsToKill=()
@@ -3416,11 +3421,18 @@ function Rethrow()
 			# whose SIGUSR2 trap handler will install a DEBUG handler that will skip simple commands until it finds a "Catch".
 			# Some bash commands are interuptable by SIGINT but if we are running a bash script, all the parents between us and
 			# tryStatePID must be bash subshells stopped on bash functions.
-			[ "$_ae_traceCatchFlag" ] && bgtrace "!!! kill -SIGUSR2 $tryStatePID  "
+			if [ "$traceCatchFlag" ]; then
+				bgtrace "!!!throwing exception that will be caught"
+				[ "$throwingStatePID" != "$$" ] && bgtrace "   PID of throw ='$throwingStatePID'"
+				[ "$throwingStatePID" != "tryStatePID" ] && bgtrace "   PID of catch ='$tryStatePID'"
+				bgtracePSTree
+				declare -g traceCatchFlag="1"
+			fi
+
 			kill -SIGUSR2 "$tryStatePID"     # this wont return if we are tryStatePID
 			(( ${#pidsToKill[@]} > 0 )) && kill -SIGINT "${pidsToKill[@]}"
-			[ "$tryStatePID" != "$BASHPID" ] && bgExit  ${_ae_exitCode:-36}
-			[ "$_ae_traceCatchFlag" ] && bgtrace "!!! returning normally. this is an error"
+			[ "$tryStatePID" != "$throwingStatePID" ] && bgExit  ${_ae_exitCode:-36}
+			[ "$traceCatchFlag" ] && bgtrace "!!! assertError: this exception is being caught but assertError is ending normally. this is an error"
 			;;
 
 		*)	echo "error: logic error. In assertError the action was computed to be '$tryStateAction' but should be one of catch,abort,continue,exitOneShell"
@@ -3517,7 +3529,7 @@ function Try()
 {
 	local funcDepthOffset=1 traceCatchFlag
 	while true; do case $1 in
-		--traceCatch)   traceCatchFlag='bgtrace "!!!$FUNCNAME | $BASH_COMMAND"' ;;
+		--traceCatch)   traceCatchFlag="--traceCatchFlag" ;;
 		--decFuncDepth) funcDepthOffset=2 ;;
 		 *)  bgOptionsEndLoop "$@" && break; set -- "${bgOptionsExpandedOpts[@]}"; esac; shift;
 	done
@@ -3535,27 +3547,32 @@ function Try()
 	local tryStateTryStatementLocation; local -A stackFrame=(); bgStackFrameGet "$funcDepthOffset" stackFrame; tryStateTryStatementLocation="${stackFrame[cmdLoc]}"
 	local tryStateIFS="$IFS"
 	local tryStateExtdebug="$(shopt -p extdebug)"
-	local debugTrapScript='bgBASH_debugTrapLINENO=$((LINENO))
-		'"${traceCatchFlag}"'
+	local debugTrapScript=''"${traceCatchFlag:+echo \"########## ENTER UNWIND TRAP ##########(\${FUNCNAME:-main\}) | \$BASH_COMMAND\" >>$_bgtraceFile;}"':
 		if (( ${#BASH_SOURCE[@]} < '"$tryStateFuncDepth"' )); then
-			IFS="$bgWS" # no need to save because we will restore the tryStateIFS copy when we return to user code
+			IFS="$bgWS" # no need to save because we are ending
 			assertError --critical "For Try block located at '"$tryStateTryStatementLocation"' no Catch block was found in the same Function. Check that code in the Try block did not skip the Catch by returning\n" >&2
 
 		elif (( ${#BASH_SOURCE[@]} > '"$tryStateFuncDepth"' )); then
-			(exit 2) # set exit code to simulate a return
+			# in ubuntu 22.04 (bash5.1) returning 2 resulted in an infinite loop entering this DEBUG trap over and over.
+			#setExitCode 2 # set exit code to simulate a return
+			setExitCode 1 # set exit code to simulate a return
 
 		elif  [[ ! "$BASH_COMMAND" =~ ^Catch:?([[:space:]]|$) ]]; then
-			(exit 1) # set exit code to not run BASH_COMMAND, go to the next command
+			setExitCode 1 # set exit code to not run BASH_COMMAND, go to the next command
 
 		else # bingo. BASH_COMMAND == Catch: at this funcDepth
-			IFS="$bgWS" # no need to save because we will restore the tryStateIFS copy when we return to user code
+			'"${traceCatchFlag:+set +x}"'
+			IFS="$bgWS" # no need to save because we will restore the tryStateIFS copy below
+
+			# restore the debugger DEBUG trap if it was installed ehn we started unwining
+			# we are in a debug trap so this line wont interrupt us and the rest of the function will complete
 			bgTrapStack pop DEBUG
 
 			unset bgBASH_debugTrapLINENO
 			IFS="'"$tryStateIFS"'" # return IFS to the value it had at the try statement
 			'"$tryStateExtdebug"'  # return the extdebug shopt to the value it had in at the try statement
 			bgBASH_tryStackWasThrown[0]="1" # the Catch function will check this to know the context its being called in
-			(exit 0) # run BASH_COMMAND and since we restore the DEBUG trap, the script will resume from there
+			setExitCode 0 # run BASH_COMMAND and since we restore the DEBUG trap, the script will resume from there
 		fi
 	'
 
@@ -3571,11 +3588,20 @@ function Try()
 	bgBASH_tryStackWasThrown=(""                   "${bgBASH_tryStackWasThrown[@]}")
 
 	# install our try block trap
-	bgTrapStack push SIGUSR2 '
+	bgTrapStack push SIGUSR2 \
+		'bgDebuggerGlobalDisable=1
+		if [ "$traceCatchFlag" ]; then
+			echo "########## ENTER SIGUSR TRAP ##########" >>$_bgtraceFile
+			exec {BASH_XTRACEFD}>>$_bgtraceFile
+			# note: that the first char of PS4 will be repeated to reflect the function depth
+			export PS4="+(${BASH_SOURCE##*/}:${LINENO}): ${FUNCNAME[0]:-main}() | "
+			set -x
+		fi
 		builtin trap - SIGUSR2
-		bgTrapStack push DEBUG '\'''"$debugTrapScript"''\''
 		shopt -s extdebug
 		set +o errtrace # extdebug turns this on but unit tests need it off
+		'"${traceCatchFlag:+bgtrace !!! SIGUSR2 removed this SIGUSR2 trap. the next line should install and switch to the DEBUG trap }"'
+		bgTrapStack push DEBUG '\''bgDebuggerGlobalDisable=; '"$debugTrapScript"''\''
 	'
 	[ "$traceCatchFlag" ] && bgtrace "!!! Try: installed SIGUSR2 in pid='$BASHPID'"
 	return 0
@@ -4164,7 +4190,7 @@ function fsTouch()
 	fi
 
 	if [ "$userOwner$groupOwner" ]; then
-		# note that this blocks works if one or the other is empty
+		# note that this blocks works even if one or the other is empty
 		if [[ ! "$curMode" =~ ^${userOwner:-[^:]*}:${groupOwner:-[^:]*}:.*$ ]]; then
 			[ "$checkOnlyFlag" ] && return 6
 			bgsudo "${sudoPrompt[@]}" --chown "$fileOrFolder:$userOwner:$groupOwner" chown $userOwner:$groupOwner "$fileOrFolder"  || assertError
